refactor: move libp2p test utilities
